[docs] Added GitFlow document (#33)
In the path '/docs/source code management' of the project, there are two new files called GitFlow in which there is all the information needed by a developer in order to perform a commit.

This can also be found in the [wiki](https://github.com/AIR-FOI-HR/AIR19E1/wiki) of the GitHub repository.