[docs] Added Third party services document (#30)
In the path '/docs/third party services' of the project, there are two new files called Third Party Services in which there is all the information about what other services we use and on what purpose.

This can also be found in the wiki (https://github.com/AIR-FOI-HR/AIR19E1/wiki) of the GitHub repository.