[docs] Added a three layer architectural diagram (#28)
In the path '/docs/three layer architectural diagram' of the project, there is a new file called Three Layer Architectural Diagram in which it is explained how the application is supposed to work on the insides.

This can also be found in the wiki (https://github.com/AIR-FOI-HR/AIR19E1/wiki) of the GitHub repository.