[docs] Added Modularity documentation (#29)
In the path '/docs/modularity' of the project, there is a new file called Modularity in which it is explained how the application is supposed to work on the insides at a module level.

This can also be found in the wiki (https://github.com/AIR-FOI-HR/AIR19E1/wiki) of the GitHub repository.